requirements.txt
bleach==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.2
coverage==7.11.0
cycler==0.12.1
distlib==0.3.9
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.3.2
idna==3.10
Jinja2==3.1.6
kaggle==1.7.4.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pandas==2.3.1
pillow==11.2.1
platformdirs==4.3.8
protobuf==6.32.0
pycodestyle==2.14.0
pyparsing==3.2.3
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
requests==2.32.5
setuptools==80.7.1
six==1.17.0
sympy==1.14.0
text-unidecode==1.3
torch==2.7.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.31.2
webencodings==0.5.1
